feat: Modified Toggle button
