+++ src/components/Header.jsx
@@ -1,75 +1,53 @@
-import React from 'react';
+import React, { useState } from 'react';
 
-export default function Navbar() {
-    const [isMenuOpen, setIsMenuOpen] = useState(false);
-    const [isDarkMode, setIsDarkMode] = useState(false);
-    const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
-  
-  // Load the theme from Local storage Using use state effect
-  // Load theme from localStorage on mount
-  useEffect(() => {
-      const storedTheme = localStorage.getItem('theme');
-      if (storedTheme === 'dark') {
-        setIsDarkMode(true);
-        document.documentElement.classList.add('dark');
-      }
-    }, []);
+export default function Navbar({ isDarkMode, toggleDarkMode }) {
+  const [isMenuOpen, setIsMenuOpen] = useState(false);
+  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
 
-    
-//  for dark mode changes
-// It will add dark mode changes
-  useEffect(() => {
-    if (isDarkMode) {
-      localStorage.setItem('theme', 'dark');
-      document.documentElement.classList.add('dark');
-    } else {
-      localStorage.setItem('theme', 'light');
-      document.documentElement.classList.remove('dark');
-    }
-  }, [isDarkMode]);
-
-  return(
+  return (
     <header className='sticky top-0 z-50 bg-[#0B1517] dark:bg-gray-900 shadow-md'>
-        <nav className="max-w-7xl mx-auto px-4 py-3 flex justify-between items-center">
+      <nav className="max-w-7xl mx-auto px-4 py-3 flex justify-between items-center">
         <button
-    className="md:hidden text-white text-3xl"
-    onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
-  >
-    ☰
-  </button>
-  <ul className={`flex-col md:flex-row absolute md:static top-16 left-0 w-full md:w-auto bg-[#0B1517] md:bg-transparent z-40 md:flex items-start md:items-center text-white transition-all duration-300 ease-in-out ${isMobileMenuOpen ? 'flex' : 'hidden'} md:gap-6 gap-4 px-4 py-4 md:py-0`}>
-  <li><a href="#home" className='hover:text-white text-[#C99A56] font-roboto'>Home</a></li>
-  <li><a href="#feedback" className='hover:text-white text-[#C99A56] font-roboto'>Feedback</a></li>
-  {/* Menu Dropdown with React State */}
-  <li
-    className="relative"
-    onMouseEnter={() => setIsMenuOpen(true)}
-    onMouseLeave={() => setIsMenuOpen(false)}
-  >
-    <div className="flex items-center gap-1 cursor-pointer text-[#C99A56] hover:text-white font-roboto">
-      <span>Menu</span>
-      <span className="text-xs">▼</span>
-    </div>
+          className="md:hidden text-white text-3xl"
+          onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
+        >
+          ☰
+        </button>
 
-    {isMenuOpen && (
-      <div className="absolute top-full left-0 mt-2 bg-white dark:bg-gray-800 text-black dark:text-white shadow-lg rounded-md p-2 min-w-[150px] z-50 transition-all duration-300">
-        <a href="#starters" className="block px-3 py-1 hover:text-black text-[#C99A56] font-roboto">Starters</a>
-        <a href="#main-course" className="block px-3 py-1 hover:text-black text-[#C99A56] font-roboto">Main Course</a>
-        <a href="#desserts" className="block px-3 py-1 hover:text-black text-[#C99A56] font-roboto">Desserts</a>
-        <a href="#drinks" className="block px-3 py-1 hover:text-black text-[#C99A56] font-roboto">Drinks</a>
-      </div>
-    )}
-  </li>
-</ul>
+        <ul className={`flex-col md:flex-row absolute md:static top-16 left-0 w-full md:w-auto bg-[#0B1517] md:bg-transparent z-40 md:flex items-start md:items-center text-white transition-all duration-300 ease-in-out ${isMobileMenuOpen ? 'flex' : 'hidden'} md:gap-6 gap-4 px-4 py-4 md:py-0`}>
+          <li><a href="#home" className='hover:text-white text-[#C99A56] font-roboto'>Home</a></li>
+          <li><a href="#feedback" className='hover:text-white text-[#C99A56] font-roboto'>Feedback</a></li>
 
-
-            <div className='flex-shrink-0'>
-                <img src="/public/images/Actions.png" alt="Meskot-Logo" className='h-10 object-contain' />
-
+          {/* Dropdown  Menu*/}
+          <li
+            className="relative"
+            onMouseEnter={() => setIsMenuOpen(true)}
+            onMouseLeave={() => setIsMenuOpen(false)}
+          >
+            <div className="flex items-center gap-1 cursor-pointer text-[#C99A56] hover:text-white font-roboto">
+              <span>Menu</span>
+              <span className="text-xs">▼</span>
             </div>
-            <div className="flex items-center gap-4">
+
+            {isMenuOpen && (
+              <div className="absolute top-full left-0 mt-2 bg-white dark:bg-gray-800 text-black dark:text-white shadow-lg rounded-md p-2 min-w-[150px] z-50 transition-all duration-300">
+                <a href="#starters" className="block px-3 py-1 hover:text-black text-[#C99A56] font-roboto">Starters</a>
+                <a href="#main-course" className="block px-3 py-1 hover:text-black text-[#C99A56] font-roboto">Main Course</a>
+                <a href="#desserts" className="block px-3 py-1 hover:text-black text-[#C99A56] font-roboto">Desserts</a>
+                <a href="#drinks" className="block px-3 py-1 hover:text-black text-[#C99A56] font-roboto">Drinks</a>
+              </div>
+            )}
+          </li>
+        </ul>
+
+        <div className='flex-shrink-0'>
+          <img src="/public/images/Actions.png" alt="Meskot-Logo" className='h-10 object-contain' />
+        </div>
+
+        <div className="flex items-center gap-4">
+          {/* Toggle Theme Button */}
           <button
-            onClick={() => setIsDarkMode(!isDarkMode)}
+            onClick={toggleDarkMode}
             className="bg-gray-700 text-white px-3 py-1 rounded-md hover:bg-gray-600 text-sm"
           >
             {isDarkMode ? '☀️ Light' : '🌙 Dark'}
@@ -82,7 +60,7 @@ export default function Navbar() {
             Menu
           </a>
         </div>
-        </nav>
+      </nav>
     </header>
-)
+  );
 }
